+++ packages/rocketh-export/CHANGELOG.md
@@ -1,5 +1,13 @@
 # rocketh-export
 
+## 1.0.2
+
+### Patch Changes
+
+- use .chain file and include genesisHash
+- Updated dependencies
+  - rocketh@0.7.2
+
 ## 1.0.1
 
 ### Patch Changes