+++ packages/rocketh-verifier/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-verifier
 
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.7.2
+
 ## 0.7.1
 
 ### Patch Changes